requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
requests>=2.31.0

# AI & RAG
llama-index>=0.10.30
llama-index-llms-groq>=0.1.3
llama-index-llms-gemini>=0.1.6
llama-index-vector-stores-supabase>=0.1.1
llama-index-embeddings-huggingface>=0.2.0

# PDF Processing
pdf2image>=1.17.0
# marker-pdf # Uncomment if using Marker (requires system dependencies)
PyMuPDF>=1.24.0

# Database & Storage
supabase>=2.4.0
redis>=5.0.3
celery>=5.3.6
flower>=2.0.0
